% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{allDrugCombinations}
\alias{allDrugCombinations}
\title{Get all drug combinations present in a MySQL database}
\usage{
allDrugCombinations(CON)
}
\arguments{
\item{CON}{MySQLConnection object}
}
\value{
list where first element are the names of all drug combinations in the database
}
\description{
Get all drug combinations present in a MySQL database
}
\examples{
# not yet
}

