% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{allModes}
\alias{allModes}
\title{Generate character vector of all possible interaction mode classifications
implemented in classifyByThOutcomeVector()}
\usage{
allModes()
}
\value{
MODES character vector of interaction mode classifications. See
the vignettes on the implementation of mode classification.
}
\description{
Generate character vector of all possible interaction mode classifications
implemented in classifyByThOutcomeVector()
}
\examples{
MODES <- allModes()
}

