% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{asCountsMatrix}
\alias{asCountsMatrix}
\title{Strip counts data.frame of any extra variables}
\usage{
asCountsMatrix(counts)
}
\arguments{
\item{counts}{matrix of counts where each row is a gene and each column is a library}
}
\value{
counts matrix of where each row is a gene and each column is a library
}
\description{
Strip counts data.frame of any extra variables
}

