% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classification.R
\name{classifyByThOutcomeVector}
\alias{classifyByThOutcomeVector}
\title{Classify a gene into an interaction mode given an outcome vector and log-fold changes}
\usage{
classifyByThOutcomeVector(ov_logfc, pipeline = "limma")
}
\arguments{
\item{ov_logfc}{vector of length 1:6 + n, where elements 1:6 are the elements of the pairwise comparison outcome vector,
n is a positive integer that satisfies n %% 2 == 0 and elements 6 : 6 + n are log fold changes associated with
each gene}

\item{character}{vector indicating which pipeline should be used to code outcome vectors}
}
\value{
character vector of length 1 indicating interaction mode or lack thereof (null vector)
}
\description{
Classify a gene into an interaction mode given an outcome vector and log-fold changes
}

