% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/limma_generics.R
\name{codeLimmaOutcomeVectors}
\alias{codeLimmaOutcomeVectors}
\title{code pairwise comparison outcome vectors with limma}
\usage{
codeLimmaOutcomeVectors(eb, design)
}
\arguments{
\item{eb}{limma::eBayes()}

\item{design}{design matrix initialized with model.matrix}

\item{DRUG_COMBINATION}{Boolean TRUE when conditions of interaction are vehicle, signal
x, signal y, and signal x + y respectively}
}
\value{
matrix where matrix[,1:6] are the six elements of the outcome vectors and
matrix[,7:10] are the log fold changes from limma::topTable()
}
\description{
code pairwise comparison outcome vectors with limma
}
\examples{
library(limma)
}

