% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{codeOVElement}
\alias{codeOVElement}
\title{Code elements of pairwise outcome vectors with confidence intervals from limma}
\usage{
codeOVElement(condition1, condition2)
}
\arguments{
\item{condition1}{matrix where conditions 1:3 are the log fold changes}

\item{condition2}{""}
}
\value{
numeric vector of length n where each element represents the outcome of the ith
pairwise comparison
}
\description{
Code elements of pairwise outcome vectors with confidence intervals from limma
}

