% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{comboExists}
\alias{comboExists}
\title{return boolean indicating if drug combination was found in the database}
\usage{
comboExists(combo.tbl.index)
}
\arguments{
\item{combo.tbl.index}{numeric vector indicating vector indices of signal combination
in library frequency distribution table}
}
\value{
boolean indicating if the combination exists in the database
}
\description{
return boolean indicating if drug combination was found in the database
}

