% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{comboIndices}
\alias{comboIndices}
\title{return signal combination indices in treatment/concentration table}
\usage{
comboIndices(T1, T2, treatment.conc.tbl, conc)
}
\arguments{
\item{T1}{character vector of length 1 indicating Signal X}

\item{T2}{character vector of length 1 indicating Signal Y}

\item{treatment.conc.tbl}{table (1d vector)}
}
\value{
vector indices in table where Signal X/Signal Y combination
is present
}
\description{
return signal combination indices in treatment/concentration table
}
\examples{
\dontrun{
comboIndices("Atorvastatin", "Pioglitazone", treatment_conc_tbl)
}
}

