% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.derivative.R
\name{compute.derivative}
\alias{compute.derivative}
\title{Compute Derivative function}
\usage{
compute.derivative(varcov, diffTerm)
}
\arguments{
\item{varcov}{the variance-covariance matrix from any model of interest}

\item{diffTerm}{the variable in the model with respect to which to partial derivatives of the regression equation are computed}
}
\value{
partial derivative of the regression equation
}
\description{
This function finds partial derivatives of the regression equation with respect a given variable in the model.
}
\examples{

dat <- read.csv(system.file("example.csv", package="Interactions"))
reg <- lm(polity2 ~ Exclusivity_Index * time_since_initial_const, data = dat)
compute.derivative(vcov(reg), "Exclusivity_Index")
}
