% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeR_generics.R
\name{designColnames}
\alias{designColnames}
\title{check column names of design matrix}
\usage{
designColnames(design)
}
\arguments{
\item{design}{design matrix initialized with model.matrix}
}
\value{
design matrix with validated column names
}
\description{
check column names of design matrix
}
\examples{
\dontrun{
this_design <- designMatrixAndContrasts(metadata, drug_combination = TRUE,
pipeline = "edgeR")
this_design <- designColnames(design)
}
}

