% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeR_generics.R
\name{designMatrixAndContrasts}
\alias{designMatrixAndContrasts}
\title{make design matrix and contrasts based on whether signal_combination is 2 drugs or M/F/Dug}
\usage{
designMatrixAndContrasts(metadata, pipeline, DRUG_COMBINATION = TRUE)
}
\arguments{
\item{metadata}{data.frame with metadata retrieved for 4 conditions}

\item{pipeline}{character vector of length 1 indicating what pipeline is
being used to classify interaction modes for each gene}

\item{drug_combination}{Boolean TRUE when signal x, signal y, and signal x + y
are conditions 2, 3, and 4 respectively}
}
\value{
list where list[[1]] is the design matrix and list[[2]]
represent the contrasts created with limma::make.contrasts if pipeline == "edgeR"
if pipeline == "limma", then a design matrix is returned.
}
\description{
make design matrix and contrasts based on whether signal_combination is 2 drugs or M/F/Dug
}

