% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dge2voom_q.R
\name{dge2voom_q}
\alias{dge2voom_q}
\title{calculate normalization factors with edgeR and voom with limma to model as
approximately normal}
\usage{
dge2voom_q(counts, genes, null = FALSE)
}
\arguments{
\item{counts}{numeric matrix of counts where rows are genes and columns are libraries}

\item{genes}{data.frame of gene metadata}

\item{null}{Boolean TRUE if columns of counts matrix should be permuted to
simulate null distribution}
}
\value{
\code{d} DGEList with normalization factors calculated and
\code{voom(d)} is returned locally
}
\description{
calculate normalization factors with edgeR and voom with limma to model as
approximately normal
}
\examples{
\dontrun{
# v <- dge2voom_q(counts, genes)
# d <- estimateCommonDisp
}
}

