% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classification.R
\name{edgeRModeClassifier}
\alias{edgeRModeClassifier}
\title{Interaction mode classification with edgeR likelihood ratio tests}
\usage{
edgeRModeClassifier(counts, genes, metadata, design, alpha = 0.01,
  DRUG_COMBINATION = TRUE)
}
\arguments{
\item{counts}{numeric matrix of counts where rows are genes and columns are libraries}

\item{genes}{data.frame of gene metadata}

\item{metadata}{data.frame with metadata retrieved for 4 conditions}

\item{alpha}{numeric vector indicating what alpha should be used to
code pairwise comparison outcome vectors}

\item{null}{Boolean TRUE if null distribution should be simulated by permuting columns
of counts matrix}
}
\value{
list of interaction modes & classes, outcome vectors, EList (voom), lmFit() object,
eBayes() object, design matrix and summary table
}
\description{
Interaction mode classification with edgeR likelihood ratio tests
}

