% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{enumerateThOutcomeVectors}
\alias{enumerateThOutcomeVectors}
\title{Enumerate theoretical pairwise comparison outcome vectors given n conditions and k ranks}
\usage{
enumerateThOutcomeVectors()
}
\value{
matrix where each row represents a pairwise comparison outcome vector
}
\description{
Enumerate theoretical pairwise comparison outcome vectors given n conditions and k ranks
}

