% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeR_generics.R
\name{estimateAllGLMDisp}
\alias{estimateAllGLMDisp}
\title{Fit generalized linear model and make pairwise comparisons based on contrasts}
\usage{
estimateAllGLMDisp(d, design)
}
\arguments{
\item{d}{DGEList object with normalization factors calculated}

\item{design}{design matrix initialized with model.matrix}
}
\value{
d DGEList with common, trended and tagwise dispersion estimated
}
\description{
Fit generalized linear model and make pairwise comparisons based on contrasts
}
\examples{
# dge2voom_q returns d, a DGEList with normalization factors
# calculated to the global environment
\dontrun{v <- dge2voom_q(counts, genes)
d <- estimateAllGLMDisp(d, design)
}
}

