% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{filterMetadata}
\alias{filterMetadata}
\title{Filter data.frame of libraries retrieved by attributes such as concentration,
cell type, and donor sex}
\usage{
filterMetadata(this_combos, conc, sex, cell.type)
}
\arguments{
\item{conc}{character vector indicating concentration at which the drug was a
dministered (currently supports {"Low", "Mid", "High", "Super", "Cmax"})}

\item{sex}{character vector indicating donor sex ("M" or "F")}

\item{cell.type}{character vector of length 1 indicating cell type}
}
\value{
filtered data.frame
}
\description{
Filter data.frame of libraries retrieved by attributes such as concentration,
cell type, and donor sex
}

