% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{getCondition}
\alias{getCondition}
\title{Retrieve libraries treated with a particular condition and filter by various parameters}
\usage{
getCondition(CON, condition, conc = "", sex = "", cell.type = "")
}
\arguments{
\item{CON}{MySQLConnection object}

\item{condition}{treatment_name in treatment table. If searching for a drug
combination, condition should be a character vector of length n > 1.}

\item{conc}{character vector of length 1 indicating concentration of drug
delivered.}

\item{sex}{character vector of length 1 indicating sex of donors to be retrieved}

\item{cell.type}{character vector of length 1 indicating cell type of donors}
}
\value{
data.frame with each row representing a library that satisfies all conditions
specified in parameters
}
\description{
Retrieve libraries treated with a particular condition and filter by various parameters
}
\examples{
# not yet
}

