% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{getExpectedCounts}
\alias{getExpectedCounts}
\title{Get matrix of expected counts from rnaseq_assay_id}
\usage{
getExpectedCounts(CON, metadata, level = "gene")
}
\arguments{
\item{CON}{MySQLConnection object (should be connected to database with
table "gene_rsem_count" containing field "rnaseq_assay_id")}

\item{metadata}{data.frame where rows are libraries and columns are
variables associated with libraries}

\item{level}{character vector of length n = 1 indicating if "gene" or
"isoform" expected counts should be used. NOTE: ONLY SUPPORTS "gene" argument
for now}
}
\value{
matrix where each row is a gene and each column is a library
}
\description{
Get matrix of expected counts from rnaseq_assay_id
}

