% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{getGeneMetadata}
\alias{getGeneMetadata}
\title{Get data.frame with gene metadata}
\usage{
getGeneMetadata(CON, counts, level = "gene")
}
\arguments{
\item{CON}{MySQLConnection object}

\item{counts}{matrix of counts where rows are genes and columns are libraries}

\item{level}{character vector indicating whether gene or isoform counts should be retrieved. NOTE:
only "gene" is supported as of now.}
}
\value{
data.frame with gene metadata
}
\description{
Get data.frame with gene metadata
}

