% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_go.R
\name{getGeneVector}
\alias{getGeneVector}
\title{Get binary factor vector for genes}
\usage{
getGeneVector(genes, interaction.tbl, filter.mode = FALSE)
}
\arguments{
\item{genes}{data.frame of gene metadata (data$voom or data$dgelist)}

\item{interaction.tbl}{gene interaction table constructed from genes classified into
an interaction mode}

\item{filter.mode}{Boolean FALSE when all genes in the interaction table are used
to construct the gene vector. TRUE when the gene vector should be constructed only
from genes classified into a particular mode}
}
\value{
factor vector where genes of interest are represented by 1 and the opposite
are represented by 0
}
\description{
Get binary factor vector for genes
}
\examples{
# construct gene vector from genes exhibiting high stabilization mode
\dontrun{
this_gene_vector <- getGeneVector(data$dgelist$genes, edgeR_interaction.tbl,
filter.mode = "High.Stab")
}
}

