% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_go.R
\name{getGOTable}
\alias{getGOTable}
\title{Generate gene ontology table and data with topGO from interaction data.frame
using Fisher's Exact Test}
\usage{
getGOTable(data, interaction.tbl, ontology, filter.mode = FALSE)
}
\arguments{
\item{data}{EList or DGEList with massaged counts and all 0 rows removed
(see dge2voom_q())}

\item{interaction.tbl}{data.frame where rows are genes and
(see interactionGeneTable())}

\item{ontology}{character vector of length n = 1 indicating gene ontology}

\item{filter.mode}{Boolean FALSE when all genes in the interaction table are used
to construct the gene vector. TRUE when the gene vector should be constructed only
from genes classified into a particular mode}
}
\value{
list where list[[1]] is a topGO table and list[[2]] is a topGO data object
}
\description{
Generate gene ontology table and data with topGO from interaction data.frame
using Fisher's Exact Test
}
\examples{
# not really
}

