% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_go.R
\name{getGOTermsForAllModes}
\alias{getGOTermsForAllModes}
\title{get GO tables for all ontologies, given an interaction table data.frame constructed
from geneInteractionTable()}
\usage{
getGOTermsForAllModes(data, interaction.tbl)
}
\arguments{
\item{interaction.tbl}{data.frame constructed with geneInteractionTable()}

\item{v}{EList constructed with voom()}
}
\value{
nested list where each element is a list containing GO tables
and significant genesgenerated with topGO using each mode as a gene vector
}
\description{
get GO tables for all ontologies, given an interaction table data.frame constructed
from geneInteractionTable()
}
\examples{
\dontrun{
v <- dge2voom_q(counts, genes)
interactionModeClassification()
all_go_terms <- getGOTermsForModesAll(v, edgeR_interaction.tbl)
}
}

