% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_go.R
\name{getGOTermsForModes}
\alias{getGOTermsForModes}
\title{get topGO tables and significant genes for a particular ontology with gene
vectors for every interaction mode present in the interaction table}
\usage{
getGOTermsForModes(data, interaction.tbl, ontology)
}
\arguments{
\item{interaction.tbl}{interaction data.frame generated from
interaction mode classification}

\item{ontology}{character vector of length 1 indicating ontology ("MF", "CC", "BP")}

\item{v}{Elist constructed with dge2voom_q()}
}
\value{
list where first element is a topGO table and the second element is a
character vector of significant genes
}
\description{
get topGO tables and significant genes for a particular ontology with gene
vectors for every interaction mode present in the interaction table
}
\examples{
\dontrun{
v <- dge2voom_q(counts, genes)
interactionModeClassification()
mf_terms <- getGOTermsForModes(v, edgeR_interaction.tbl, ontology = "MF")
}
}

