% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{getIntClass}
\alias{getIntClass}
\title{Get character string classification for interaction class}
\usage{
getIntClass(logfc, pipeline = "limma")
}
\arguments{
\item{logfc}{Numeric vector of log fold changes or character vector indicating interaction mode}

\item{pipeline}{Character vector ("limma" or "edgeR")}
}
\value{
The interaction class of \code{logfc}
}
\description{
Get character string classification for interaction class
}

