% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{getInteraction}
\alias{getInteraction}
\title{Get libraries for a specified signal interaction}
\usage{
getInteraction(CON, T1, T2, T12, cell.type)
}
\arguments{
\item{CON}{MySQLConnection}

\item{T1}{Signal X, character vector of length 1 corresponding to treatment name in the treatment table}

\item{T2}{Signal Y, character vector of length 1 corresponding to treatment name in the treatment table}

\item{T12}{Signal X + Y, character vector of length 1 corresponding to treatment name for signal
combination in treatment table}

\item{cell.type}{Character vector of length 1 indicating cell type of interest}
}
\value{
data.frame for libraries of interest where rows are libraries and columns are variables associated
with those libraries
}
\description{
Get libraries for a specified signal interaction
}

