% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret_knn.R
\name{getPCAScores}
\alias{getPCAScores}
\title{Principal Component Analysis scores using log-fold changes as feature vectors}
\usage{
getPCAScores(pooled_lfc)
}
\arguments{
\item{pooled_lfc}{matrix of log fold change ratios for pairwise comparisons.
\code{pooled_lfc} is intended to be a matrix of pooled log fold change ratios that map
to known interaction modes}
}
\value{
matrix of PCA scores
}
\description{
Principal Component Analysis scores using log-fold changes as feature vectors
}
\examples{
\dontrun{
#FINISH LATER
}
}

