% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{getVehicles}
\alias{getVehicles}
\title{Get vehicles for libraries of interest}
\usage{
getVehicles(CON, cell.type, metadata)
}
\arguments{
\item{CON}{MySQLConnection object}

\item{cell.type}{Character vector indicating which cell type to retrieve}

\item{metadata}{data.frame with metadata retrieved for 4 conditions}
}
\value{
data.frame with libraries from \code{metadata} and respective vehicles
}
\description{
Get vehicles for libraries of interest
}

