% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeR_generics.R
\name{glmFitAndLRT}
\alias{glmFitAndLRT}
\title{generalized linear model and likelihood ratio tests with edgeR}
\usage{
glmFitAndLRT(d, design, contrasts)
}
\arguments{
\item{d}{DGEList with dispersions estimated}

\item{design}{list where design[[1]] is the design matrix constructed by model.matrix
and design[[2]] are the contrasts constructed by make.contrasts}

\item{contrasts}{optional constrasts constructed by make.contrasts}
}
\value{
pairwise list of likelihood ratio test tables generated with
edgeR::glmLRT
}
\description{
generalized linear model and likelihood ratio tests with edgeR
}
\examples{
\dontrun{
d <- dge2voom_q(counts, genes)
design <- designMatrixAndContrasts(metadata, drug_combination = TRUE,
pipeline = "edgeR")
my_lrt_res <- glmFitAndLRT(d, design)
}
}

