% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interaction.R
\name{initInteraction}
\alias{initInteraction}
\title{initalize an Interaction object from a list generated with the limma or edge classifier
functions}
\usage{
initInteraction(l)
}
\arguments{
\item{l}{list with elements seen in the function definition.}
}
\value{
Interaction object
}
\description{
initalize an Interaction object from a list generated with the limma or edge classifier
functions
}

