% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{interactionDistTable}
\alias{interactionDistTable}
\title{Generate table with outcome vectors, interaction classes & modes for all genes}
\usage{
interactionDistTable(outcome.vectors, modes, classes)
}
\arguments{
\item{outcome.vectors}{matrix where nrow(matrix) == n or character vector of length = n indicating outcome vectors}

\item{modes}{character vector of length n indicating interaction modes}

\item{classes}{character vector of length n indicating interaction classes}
}
\value{
data.frame with outcome vectors, interaction modes, and classes for each gene
}
\description{
Generate table with outcome vectors, interaction classes & modes for all genes
}

