% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{interactionGeneTable}
\alias{interactionGeneTable}
\title{Generate table with genes classified into a mathematically defined interaction mode}
\usage{
interactionGeneTable(v, modes, log.fcs, pw_s, classes,
  exclude.null.vector = TRUE, exclusivity.vectors = FALSE,
  DRUG_COMBINATION = TRUE, ...)
}
\arguments{
\item{v}{EList constructed with \code{voom}}

\item{modes}{Character vector of length n indicating interaction modes of genes}

\item{log.fcs}{Numeric matrix of log fold changes generated from the limma or edgeR pipelines}

\item{classes}{Character vector of interaction classes for n genes. If missing, getIntClass()
is called to assign interaction classes to each interaction mode.}

\item{exclude.null.vector}{Boolean TRUE when genes with the null outcome vector
(0, 0, 0, 0, 0, 0) should be excluded from the table.}

\item{exclusivity.vectors}{Boolean FALSE when exclusivity vectors should be excluded
from the table.}
}
\value{
\code{data.frame} with outcome vectors, interaction classes & modes, fold changes
for all genes classified into an interaction class
}
\description{
Generate table with genes classified into a mathematically defined interaction mode
}

