% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classification.R
\name{interactionModeClassifier}
\alias{interactionModeClassifier}
\title{Interaction mode classification with limma and edgeR}
\usage{
interactionModeClassifier(data, DRUG_COMBINATION = TRUE, ci = 0.99,
  alpha = 0.01)
}
\arguments{
\item{alpha}{numeric vector indicating what alpha should be used to
code pairwise comparison outcome vectors}

\item{counts}{numeric matrix of counts where rows are genes and columns are libraries}

\item{genes}{data.frame of gene metadata}

\item{metadata}{data.frame with metadata retrieved for 4 conditions}

\item{null}{Boolean TRUE if null distribution should be simulated by permuting
columns of counts matrix}
}
\value{
nested list
}
\description{
Interaction mode classification with limma and edgeR
}

