% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{isAnomaly}
\alias{isAnomaly}
\title{Determine if an outcome vector is enumerated theoretically}
\usage{
isAnomaly(v, th.vectors)
}
\arguments{
\item{v}{outcome vector (numeric) of length n = 6}

\item{th.vectors}{matrix where each row is an enumerated outcome vector}
}
\value{
Boolean TRUE when outcome vector is not enumerated as part of the 75 theoretical outcome vectors
}
\description{
Determine if an outcome vector is enumerated theoretically
}

