% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{isMode}
\alias{isMode}
\title{Check to see if an outcome vector is associated with a particular interaction mode}
\usage{
isMode(ov, pwt_n)
}
\arguments{
\item{ov}{numeric or character vector of length n = 1 indicating outcome vector}

\item{pwt_n}{matrix where each row is an outcome vector associated with a particular classification (mode, class)}
}
\value{
Boolean TRUE when outcome vector is associated with a particular mode
}
\description{
Check to see if an outcome vector is associated with a particular interaction mode
}

