% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{isNi}
\alias{isNi}
\title{check outcome vector to see if it is consistent with 1/5 no-interaction vectors}
\usage{
isNi(v)
}
\arguments{
\item{v}{6-dimensional outcome vector (numeric vector where length(v) == 6)}
}
\value{
Boolean TRUE if outcome vector matches 1/5 outcome vectors that are consistent
with a lack of interaction
}
\description{
check outcome vector to see if it is consistent with 1/5 no-interaction vectors
}

