% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{isValidCombination}
\alias{isValidCombination}
\title{Check to see if a drug combination exists in eld_v4}
\usage{
isValidCombination(T1, T2, DRUG_COMBINATION)
}
\arguments{
\item{T1}{Signal X can be a treatment_name in treatment table}

\item{T2}{Signal Y ""}

\item{DRUG_COMBINATION}{Boolean TRUE if T1 and T2 are both drugs}
}
\value{
A character vector of length 1 indicating signal combination is present
}
\description{
Check to see if a drug combination exists in eld_v4
}

