% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classification.R
\name{limmaModeClassifier}
\alias{limmaModeClassifier}
\title{Interaction mode classification with matrix of counts and data.frames for genes and metadata}
\usage{
limmaModeClassifier(counts, genes, metadata, design, DRUG_COMBINATION = TRUE,
  ci = 0.99, null = FALSE, ...)
}
\arguments{
\item{counts}{numeric matrix of counts where rows are genes and columns are libraries}

\item{genes}{data.frame of gene metadata}

\item{metadata}{data.frame with metadata retrieved for 4 conditions}

\item{design}{design matrix constructed with model.matrix}

\item{ci}{Confidence level for confidence intervals generated by limma}

\item{null}{Boolean TRUE if null distribution should be simulated by permuting columns
of counts matrix}

\item{drug_combination}{Boolean TRUE when signal x, signal y, and signal x + y
are conditions 2, 3, and 4 respectively}
}
\value{
list of interaction modes & classes, outcome vectors, EList (voom), lmFit() object,
eBayes() object, design matrix and summary table
}
\description{
Interaction mode classification with matrix of counts and data.frames for genes and metadata
}

