% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{massageCounts}
\alias{massageCounts}
\title{apply expected counts filter to DGEList object
Aaron Mackey}
\usage{
massageCounts(dge, cpm_min = 2, sample_min = 5)
}
\arguments{
\item{dge}{DGEList}

\item{cpm_min}{minimum counts per million}

\item{sample_min}{minimum number of samples}
}
\value{
filtered DGEList
}
\description{
apply expected counts filter to DGEList object
Aaron Mackey
}

