% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/limma_generics.R
\name{modeClassification}
\alias{modeClassification}
\title{Interaction mode classification with matrix of outcome vectors and log fold changes}
\usage{
modeClassification(res, pipeline)
}
\arguments{
\item{res}{matrix where columns 1:6 are the elements of the outcome vectors and columns
7:length(res) are the log fold changes for interaction class classification}

\item{pipeline}{character vector of length n = 1 indicating pipeline to use for
interaction mode classification ("edgeR" or "limma")}
}
\value{
character vector of length m = nrow(res) indicating interaction modees
}
\description{
Interaction mode classification with matrix of outcome vectors and log fold changes
}
\examples{
\dontrun{
library(limma)
ov <- codeLimmaOutcomeVectors(eb, design)
limma_modes <- modeClassification(ov, pipeline = "limma")
}
}

