% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{niCheck}
\alias{niCheck}
\title{Check to see if an outcome vector is consistent with a "non-interaction" mode}
\usage{
niCheck(v)
}
\arguments{
\item{v}{character vector of length n = 6 or numeric vector of length m = 1 indicating the outcome vector of interest}
}
\value{
Boolean TRUE when vector is consistent with 1/5 "non-interaction" vector
}
\description{
Check to see if an outcome vector is consistent with a "non-interaction" mode
}
\examples{
null_vector <- c(0, 0, 0, 0, 0, 0)
niCheck(null_vector)
high_stab_vector <- c(1, 1, 1, 0, 0, 0)
niCheck(high_stab_vector)
}

