% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{outcomeVectorsByMode}
\alias{outcomeVectorsByMode}
\title{Generate a list where each element is a matrix/vector of outcome vectors associated with a
known mode/class}
\usage{
outcomeVectorsByMode()
}
\value{
named list where each element contains a vector or matrix with the
outcome vector(s) associated with a particular class or mode
}
\description{
Generate a list where each element is a matrix/vector of outcome vectors associated with a
known mode/class
}
\examples{
library(magrittr)
pw <- outcomeVectorsByMode()
}

