% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{ovTable}
\alias{ovTable}
\title{generate frequency distribution table for outcome vectors}
\usage{
ovTable(pw_s, get.class = FALSE)
}
\arguments{
\item{pw_s}{Character vector of length n with pairwise comparison outcome vectors
if pw_s is a matrix, then each row will be coerced to a character vector of
length 1}
}
\value{
A data.frame with the frequency distribution of outcome vectors and
associated interaction classes
}
\description{
generate frequency distribution table for outcome vectors
}

