% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret_knn.R
\name{partitionData}
\alias{partitionData}
\title{partition feature vectors (pca_scores) and interaction mode classifications
into training and test sets for classification with KNN}
\usage{
partitionData(pooled_modes, pca_scores)
}
\arguments{
\item{pooled_modes}{character vector representing interaction mode classifications
pooled from n signal combinations}

\item{pca_scores}{matrix of PCA scores generated from princomp()}
}
\value{
list with training scores, training modes, testing scores, and testing modes
}
\description{
partition feature vectors (pca_scores) and interaction mode classifications
into training and test sets for classification with KNN
}
\examples{
\dontrun{
#FINISH LATER
}
}

