% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_scatter_matrix.R
\name{pcaScatterMatrix}
\alias{pcaScatterMatrix}
\title{Make scatter matrix from PCA scores}
\usage{
pcaScatterMatrix(modes, log.fc)
}
\arguments{
\item{modes}{character vector of length n indicating the modes of genes 1:n}

\item{log.fc}{matrix of log fold changes where ncol(log.fc) == 4 if pipeline == "limma" and
ncol(log.fc) == 6 if pipeline == "edgeR"}
}
\value{
PCA scatter matrix generated with GGally::ggpairs
}
\description{
Make scatter matrix from PCA scores
}
\examples{
library(GGally)
\dontrun{
scatter_matrix <- pcaScatterMatrix(limma_modes, logfc)
}
}

