% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_margins.R
\name{plot.sim_margins}
\alias{plot.sim_margins}
\title{Plot coefficients from simple slopes analysis}
\usage{
\method{plot}{sim_margins}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=sim_margins]{sim_margins()}} object.}

\item{...}{arguments passed to \code{\link[jtools:plot_summs]{jtools::plot_coefs()}}}
}
\description{
This creates a coefficient plot to visually summarize the
results of simple slopes analysis.
}
