% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{poolInteractionGenes}
\alias{poolInteractionGenes}
\title{Generate interaction table with genes classified into mathematically defined modes by both limma and edgeR}
\usage{
poolInteractionGenes(edgeR_interaction_tbl, limma_interaction_tbl)
}
\arguments{
\item{edgeR_interaction_tbl}{interaction table with data generated with edgeR}

\item{limma_interaction_tbl}{interaction table with data generated with limma}
}
\value{
master_interaction_tbl, interaction table with genes classified into modes by limma and edgeR
}
\description{
Generate interaction table with genes classified into mathematically defined modes by both limma and edgeR
}

