% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barplots.R
\name{qplotOutcomeVectorDist}
\alias{qplotOutcomeVectorDist}
\title{Make barplot of outcome vector distribution with base::barplot}
\usage{
qplotOutcomeVectorDist(pw_s, remove.ni = TRUE, geom_bar = TRUE, log = "")
}
\arguments{
\item{pw_s}{character vector}

\item{remove.ni}{Boolean TRUE when null outcome vector should be removed from the frequency distribution table}

\item{geom_bar}{Boolean TRUE when ggplot2 with geom_bar should be used instead of base R graphics}

\item{log}{character vector of length 1 indicating which variable (if any) to plot log of}
}
\value{
base::barplot of outcome vector frequency distribution
}
\description{
Make barplot of outcome vector distribution with base::barplot
}

