% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{rankTuples}
\alias{rankTuples}
\title{Compute row ranks and average ties (should be deprecated)}
\usage{
rankTuples(x)
}
\arguments{
\item{x}{matrix of dimensions i x j where the row ranks should be computed}
}
\value{
matrix where each row are the ranks of the original values
}
\description{
Compute row ranks and average ties (should be deprecated)
}
\examples{
library(magrittr)
rpois(16, 3) \%>\% matrix(nrow = 4, ncol = 4) \%>\% rankTuples
}

