% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{reportModeClassStats}
\alias{reportModeClassStats}
\title{Print statistics about the interaction mode distribution and return a frequency distribution
table of interaction mode classifications}
\usage{
reportModeClassStats(modes)
}
\arguments{
\item{modes}{character vector of length n indicating the interaction modes for each gene}
}
\value{
table of
}
\description{
Print statistics about the interaction mode distribution and return a frequency distribution
table of interaction mode classifications
}

