% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{reportnAnomaly}
\alias{reportnAnomaly}
\title{Report number of genes classified into an anomalous mode classifcation}
\usage{
reportnAnomaly(modes)
}
\arguments{
\item{modes}{character vector of length n indicating the interaction mode classifications
of each gene in v$genes}
}
\value{
number of genes coded into anomalous outcome vectors
}
\description{
Report number of genes classified into an anomalous mode classifcation
}
\examples{
\dontrun{
reportNAnomaly(edgeR_modes)
}
}

