% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{reportnInteractions}
\alias{reportnInteractions}
\title{Report number of genes classified into a non-null or non-anomalous interaction mode}
\usage{
reportnInteractions(modes)
}
\arguments{
\item{modes}{character vector of length n indicating the interaction modes classified
by a particular pipeline}
}
\value{
number of genes classified into a non-null and non-anomalous interaction mode
}
\description{
Report number of genes classified into a non-null or non-anomalous interaction mode
}
\examples{
\dontrun{
reportnInteractions(limma_modes)
}
}

