% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_go.R
\name{saveTopGoRes}
\alias{saveTopGoRes}
\title{save topGO table as .csv and topGO DAG as .png}
\usage{
saveTopGoRes(go.res)
}
\arguments{
\item{go.res}{list where go.res[[1]] is a topGO table and go.res[[2]] is the topGO
data object}
}
\value{
saved .csv file with topGO table in signal combination output directory
and saved .png file with topGO graph in output dir
}
\description{
save topGO table as .csv and topGO DAG as .png
}

