% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interaction_tools.R
\name{sharedGenesTable}
\alias{sharedGenesTable}
\title{Create data.frame for genes that were classified into modes by both limma and edgeR}
\usage{
sharedGenesTable(edgeR_interaction_tbl, limma_interaction_tbl)
}
\arguments{
\item{edgeR_interaction_tbl}{interaction table (data.frame) generated by edgeR
interaction mode classification}

\item{limma_interaction_tbl}{interaction table (data.frame) generated by limma
interaction mode classification}
}
\value{
interaction table (data.frame) with genes classified into non-null modes
by both limma and edgeR
}
\description{
Create data.frame for genes that were classified into modes by both limma and edgeR
}
\examples{
library(limma)
library(edgeR)
\dontrun{
master_interaction_tbl <- sharedGenesTable(edgeR_interaction_tbl,
                                         limma_interaction_tbl)
}
}

