% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_margins.R, R/simple_slopes.R
\name{glance.sim_margins}
\alias{glance.sim_margins}
\alias{tidy.sim_slopes}
\alias{glance.sim_slopes}
\title{Tidiers for \code{\link[=sim_slopes]{sim_slopes()}} objects.}
\usage{
tidy.sim_slopes(x, conf.level = 0.95, ...)

glance.sim_slopes(x, ...)
}
\arguments{
\item{x}{The \code{sim_slopes} object}

\item{...}{Ignored.}

\item{conf.level}{The width of confidence intervals. Default is .95 (95\%).}
}
\description{
You can use \code{\link[broom:tidy]{broom::tidy()}} and \code{\link[broom:glance]{broom::glance()}} for "tidy"
methods of storing \code{sim_slopes} output.
}
