\name{sitt-package}
\alias{sitt-package}
\alias{sitt}
\docType{package}
\title{
Signal Interaction Tools for Transcriptomics
~~ sitt ~~
}
\description{
 sitt provides functions for classifying gene interaction profiles
             into mathematically and biologically defined interaction modes and classes,
             extending the homogenous classes of positive and negative synergies into
             heterogenous classes comprised of defined interaction modes. This package
             currently supports transformation of RNA-seq data (expected counts)
             retrieved from the HemoShear Therapeutics eld_v4 database. Future plans
             include wrapper functions microarray data transformation ...
}
\details{
\tabular{ll}{
Package: \tab sitt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-10\cr
License: \tab Private\cr
}
see future link
}
\author{
Justin Taylor

Maintainer: Justin Taylor <taylor@hemoshear.com>
}
\references{
\emph{Combinatorial code governing cellular responses to complex stimuli}
\url{http://www.nature.com/ncomms/2015/150421/ncomms/7847/full/ncomms/7847.html}
\emph{voom: precision weights unlock linear model analysis tools for RNA-seq read counts}
\url{http://statsci.org/smyth/pubs/VoomPreprint.pdf}
\emph{edgeR: a Bioconductor pacakge for differential expression analysis of digital
gene expression data}
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2796818/}
\emph{Using transcriptional signatures to assess immune cell function: from basic mechanisms to immune-related disease}
 \url{www.sciencedirect.com/science/article/pii/S0000xxxxx}
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# not yet
}

