% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outcome_vector_tools.R
\name{stringCoerce}
\alias{stringCoerce}
\title{Coerce a numeric vector to a string the wrong way}
\usage{
stringCoerce(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
character vector of length n = 1
}
\description{
Coerce a numeric vector to a string the wrong way
}

