% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret_knn.R
\name{trainKNN}
\alias{trainKNN}
\title{train k-nearest neighbours model with pca scores as feature vectors and interaction modes
as classifications}
\usage{
trainKNN(my_train_scores, my_train_modes)
}
\arguments{
\item{my_train_scores}{matrix where each row represents a feature vector (gene)}

\item{my_train_modes}{character vector of length n where each element indicates the
interaction mode classification for a gene}
}
\value{
caret k-nearest neighbours model
}
\description{
train k-nearest neighbours model with pca scores as feature vectors and interaction modes
as classifications
}
\examples{
\dontrun{
knn_model <- trainKNN(my_pca_scores, my_modes)
}
}

