% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{treatmentConcTable}
\alias{treatmentConcTable}
\title{Generate a table of all treatment/concentration conditions for a given signal X and signal Y}
\usage{
treatmentConcTable(metadata)
}
\arguments{
\item{metadata}{data.frame of metadata retrieved from getInteraction()}
}
\value{
base::table with number of libraries retrieved for each
treatment/concentration condition
}
\description{
Generate a table of all treatment/concentration conditions for a given signal X and signal Y
}
\examples{
\dontrun{
treatmentConcTable(metadata)
}
}

