% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classification.R
\name{tuneClassifier}
\alias{tuneClassifier}
\title{Evaluate model performance by initializing many analysis objects given
a vector of alphas for coding confidence intervals}
\usage{
tuneClassifier(counts, genes, metadata, alphas, DRUG_COMBINATION = TRUE,
  null = FALSE)
}
\arguments{
\item{counts}{numeric matrix of counts where rows are genes and columns are libraries}

\item{genes}{data.frame of gene metadata}

\item{metadata}{data.frame with metadata retrieved for 4 conditions}

\item{alphas}{numeric vector of length n with alphas to code outcome vectors
with}

\item{DRUG_COMBINATION}{Boolean TRUE if drug combination}

\item{null}{Boolean TRUE if null distribution should be simulated by permuting
columns of count matrix}
}
\value{
list of n = length(alphas), see output for \code{edgeRModeClassifier}
and \code{limmaModeClassifier}
}
\description{
Evaluate model performance by initializing many analysis objects given
a vector of alphas for coding confidence intervals
}

