% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{uniqueBatches}
\alias{uniqueBatches}
\title{Get vector of unique batches retrieved from a particular database}
\usage{
uniqueBatches(metadata)
}
\arguments{
\item{metadata}{data.frame with metadata retrieved for 4 conditions}
}
\value{
vector of unique batches for a given metadata \code{data.frame}
}
\description{
Get vector of unique batches retrieved from a particular database
}

