% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{tidy.sim_slopes}
\alias{tidy.sim_slopes}
\alias{glance.sim_slopes}
\title{Tidiers for \code{\link[=sim_slopes]{sim_slopes()}} objects.}
\usage{
\method{tidy}{sim_slopes}(x, conf.level = 0.95, ...)

\method{glance}{sim_slopes}(x, ...)
}
\arguments{
\item{x}{The \code{sim_slopes} object}

\item{conf.level}{The width of confidence intervals. Default is .95 (95\%).}

\item{...}{Ignored.}
}
\description{
You can use \code{\link[broom:reexports]{broom::tidy()}} and \code{\link[broom:reexports]{broom::glance()}} for "tidy"
methods of storing \code{sim_slopes} output.
}
